rename module name
