+++ mamas-common/src/main/groovy/com/sondahum/mamas/common/model/person/Admin.groovy
@@ -1,0 +1,5 @@
+package com.sondahum.mamas.common.model.person
+
+class Admin extends User {
+    String company
+}
