+++ mamas-common/src/main/groovy/com/sondahum/mamas/common/model/person/Client.groovy
@@ -1,0 +1,5 @@
+package com.sondahum.mamas.common.model.person
+
+class Client extends User {
+
+}
