+++ mamas-common/src/main/groovy/com/sondahum/mamas/common/model/person/User.groovy
@@ -1,0 +1,7 @@
+package com.sondahum.mamas.common.model.person
+
+class User {
+    String name
+    String phone
+
+}
