+++ mamas-common/src/main/groovy/com/sondahum/mamas/common/model/product/Estate.groovy
@@ -1,0 +1,9 @@
+package com.sondahum.mamas.common.model.product
+
+import com.sondahum.mamas.common.model.type.ContractType
+import com.sondahum.mamas.common.model.type.EstateType
+
+class Estate {
+    EstateType productType
+    ContractType contractType
+}
