+++ mamas-common/src/main/groovy/com/sondahum/mamas/common/model/type/ContractType.groovy
@@ -1,4 +1,4 @@
-package com.sondahum.mamas.model.type
+package com.sondahum.mamas.common.model.type
 
 enum ContractType {
     BARGAIN(1, '매매'), // 매매
