--- mamas-common/src/main/groovy/com/sondahum/mamas/model/person/Admin.groovy
@@ -1,5 +1,0 @@
-package com.sondahum.mamas.model.person
-
-class Admin extends User {
-
-}
