--- mamas-common/src/main/groovy/com/sondahum/mamas/model/person/Client.groovy
@@ -1,8 +1,0 @@
-package com.sondahum.mamas.model.person
-
-import com.sondahum.mamas.model.type.ContractType
-import com.sondahum.mamas.model.type.EstateType
-
-class Client extends User {
-
-}
