--- mamas-common/src/main/groovy/com/sondahum/mamas/model/person/User.groovy
@@ -1,10 +1,0 @@
-package com.sondahum.mamas.model.person
-
-import com.sondahum.mamas.model.type.Role
-
-class User {
-    String name
-    String phoneNumber
-    Role role
-
-}
