--- mamas-common/src/main/groovy/com/sondahum/mamas/model/product/Estate.groovy
@@ -1,9 +1,0 @@
-package com.sondahum.mamas.model.product
-
-import com.sondahum.mamas.model.type.ContractType
-import com.sondahum.mamas.model.type.EstateType
-
-class Estate {
-    EstateType productType
-    ContractType contractType
-}
